Implemented new pattern for product/coproduct.
