+++ core/src/main/scala-3/pureconfig/generic/derivation/EnumConfigReaderDerivation.scala
